+++ storage_drivers/ontap/ontap_san.go
@@ -236,7 +236,7 @@ func (d *SANStorageDriver) Create(
 	// Get options with default fallback values
 	// see also: ontap_common.go#PopulateConfigurationDefaults
 	size := strconv.FormatUint(sizeBytes, 10)
-	spaceAllocate, _ := strconv.ParseBool(utils.GetV(opts, "spaceAllocate", d.Config.SpaceAllocate))
+	spaceAllocation, _ := strconv.ParseBool(utils.GetV(opts, "spaceAllocation", d.Config.SpaceAllocation))
 	spaceReserve := utils.GetV(opts, "spaceReserve", d.Config.SpaceReserve)
 	snapshotPolicy := utils.GetV(opts, "snapshotPolicy", d.Config.SnapshotPolicy)
 	snapshotReserve := utils.GetV(opts, "snapshotReserve", d.Config.SnapshotReserve)
@@ -277,7 +277,7 @@ func (d *SANStorageDriver) Create(
 	log.WithFields(log.Fields{
 		"name":            name,
 		"size":            size,
-		"spaceAllocate":   spaceAllocate,
+		"spaceAllocation": spaceAllocation,
 		"spaceReserve":    spaceReserve,
 		"snapshotPolicy":  snapshotPolicy,
 		"snapshotReserve": snapshotReserveInt,
@@ -310,7 +310,7 @@ func (d *SANStorageDriver) Create(
 	osType := "linux"
 
 	// Create the LUN
-	lunCreateResponse, err := d.API.LunCreate(lunPath, int(sizeBytes), osType, false, spaceAllocate)
+	lunCreateResponse, err := d.API.LunCreate(lunPath, int(sizeBytes), osType, false, spaceAllocation)
 	if err = api.GetError(lunCreateResponse, err); err != nil {
 		return fmt.Errorf("error creating LUN: %v", err)
 	}